app/requirements.txt
fastapi
uvicorn
mysql-connector-python
pyjwt
pydantic
python-dotenv
tabulate
aiomysql
reportlab
matplotlib